Fix cipher name and mode

+++ sys/geom/luks/g_luks_hmac.c
@@ -164,7 +164,7 @@ g_luks_crypto_ivgen(struct g_luks_softc *sc, off_t offset, u_char *iv,
 
 
 void
-g_luks_crypto_ivgen_aalgo(uint16_t algo, off_t offset, u_char *iv,
+g_luks_crypto_ivgen_aalgo(u_int algo, off_t offset, u_char *iv,
     size_t size)
 {
 	uint8_t off[8];
